[kimi-k2.5]vit部分做w8a8_mxfp8量化时，量化权重被强转bug

+++ vllm/model_executor/models/kimi_k25.py
@@ -375,9 +375,10 @@ class KimiK25ForConditionalGeneration(
                 quant_config=self._maybe_ignore_quant_config(quant_config),
                 prefix=maybe_prefix(prefix, "vision_tower"),
             )
-            self.vision_tower = self.vision_tower.to(
-                device=self.device, dtype=model_config.dtype
-            )
+            # self.vision_tower = self.vision_tower.to(
+            #     device=self.device, dtype=model_config.dtype
+            # )
+            self.vision_tower = self.vision_tower.to(device=self.device)
 
             self.mm_projector = KimiK25MultiModalProjector(
                 config=config.vision_config,
